+++ Particle.h
@@ -6,14 +6,13 @@
 
 class Particle {
 public:
-	Particle(int name_ptc, float m, float x, float y, float vx, float vy);
-
+	Particle(int name_ptc, double m, double x, double y, double vx, double vy);
 
 	int name;
 	int isFixed;
 
 	Vector3 getPosition(), getVelocity();
-	float getMass();
+	double getMass();
 	void print();
 	void calculate(std::vector<Particle> particles, std::vector<int> particlesInSet, std::vector<Force> forces, std::vector<int> forcesInSet, bool gravity, int t);
 	std::vector<int> getIncludedSets();
@@ -24,5 +23,5 @@ private:
 	static int count;
 	std::vector<int> includedSets;
 	Vector3 Pos, Vel, Pos_post, Vel_post;
-	float mass;
+	double mass;
 };
